fix: enable -Wextra -Werror and fix all warnings
Add stricter compiler warnings to catch potential issues early.
Fix ~25 warnings including unused parameters and sign comparisons.
Also add CLAUDE.md with build/test instructions.

+++ src/hash.c
@@ -62,7 +62,7 @@ void hashPassword(char* password, char* output) {
   }
 
   strcat(result, "/");
-  for(int i = 0; i < sizeof(salt); i++) {
+  for(size_t i = 0; i < sizeof(salt); i++) {
     sprintf(result + strlen(result), "%02x", salt[i]);
   }
 #else
